#1102 mv tests for find_or_create... into separate file and add new test for symbol-type changing.
